requirements.txt
Django==5.0.4
django-extensions==3.2.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
isort==5.13.2
mccabe==0.7.0
mysqlclient==2.2.4
pycodestyle==2.11.1 
pyflakes==3.2.0 
pytz==2024.1
six==1.16.0
python-decouple==3.8
